#9 Add specs.common for common specifications
