+++ drs4/specs/__init__.py
@@ -1,7 +1,8 @@
-__all__ = ["csv", "vdif", "zarr"]
+__all__ = ["common", "csv", "vdif", "zarr"]
 
 
 # submodules
+from . import common
 from . import csv
 from . import vdif
 from . import zarr
