+++ drs4/specs/zarr.py
@@ -3,49 +3,34 @@ __all__ = ["Zarr", "open_csvs", "open_vdifs"]
 
 # standard library
 from dataclasses import dataclass, field
-from datetime import datetime
-from typing import Literal as L, Optional, Union, get_args
+from typing import Literal as L, Optional, get_args
 
 
 # dependencies
 import numpy as np
 import xarray as xr
 from xarray_dataclasses import AsDataset, Attr, Coordof, Data, Dataof
+from .common import (
+    FREQ_INNER,
+    FREQ_OUTER,
+    Channel,
+    Chassis,
+    FreqRange,
+    Interface,
+    IntegTime,
+    SideBand,
+    Time,
+    Chan,
+    AutoUSB,
+    AutoLSB,
+    Cross2SB,
+)
 from .csv import open_csv_autos, open_csv_cross
 from .vdif import open_vdif
 from ..utils import StrPath, XarrayJoin
 
 
-# type hints
-Channel = int
-Chassis = L[1, 2]
-FreqRange = L["inner", "outer"]
-Interface = L[1, 2]
-IntegTime = L[100, 200, 500, 1000]  # ms
-SideBand = L["USB", "LSB"]
-TimeLike = Union[datetime, str]
-
-
-# constants
-CHAN_TOTAL = 512  # ch
-FREQ_INTERVAL = 0.02  # GHz
-FREQ_INNER = FREQ_INTERVAL * np.arange(CHAN_TOTAL * 0, CHAN_TOTAL * 1)  # GHz
-FREQ_OUTER = FREQ_INTERVAL * np.arange(CHAN_TOTAL * 1, CHAN_TOTAL * 2)  # GHz
-
-
 # data classes
-@dataclass
-class Time:
-    data: Data[L["time"], L["M8[ns]"]]
-    long_name: Attr[str] = "Measured time in UTC"
-
-
-@dataclass
-class Chan:
-    data: Data[L["chan"], np.int64]
-    long_name: Attr[str] = "Channel number"
-
-
 @dataclass
 class Freq:
     data: Data[L["chan"], np.float64]
@@ -63,27 +48,6 @@ class SignalChan:
 class SignalSB:
     data: Data[L["time"], L["U3"]]
     long_name: Attr[str] = "Signal sideband"
-
-
-@dataclass
-class AutoUSB:
-    data: Data[tuple[L["time"], L["chan"]], np.float64]
-    long_name: Attr[str] = "Auto-correlation spectra of USB"
-    units: Attr[str] = "Arbitrary unit"
-
-
-@dataclass
-class AutoLSB:
-    data: Data[tuple[L["time"], L["chan"]], np.float64]
-    long_name: Attr[str] = "Auto-correlation spectra of LSB"
-    units: Attr[str] = "Arbitrary unit"
-
-
-@dataclass
-class Cross2SB:
-    data: Data[tuple[L["time"], L["chan"]], np.complex128]
-    long_name: Attr[str] = "Cross-correlation spectra of 2SB"
-    units: Attr[str] = "Arbitrary unit"
 
 
 @dataclass
